added pop up menu to a filter icon
